(refactor) moved authentication handling to models -> authmodel
